Move sign up with facebook button to signup component
