+++ app/templates/signup.hbs
@@ -7,8 +7,6 @@
       Log in
     </a>
   </span>
-  
-  <button class="btn btn--blue" {{action 'signupWithFb'}}>Sign Up With Facebook</button>
 
   {{account-signup-form}}
 {{/account-wrapper}}
